Cleaned up code.

+++ src/runtime/stackframe.cpp
@@ -27,17 +27,17 @@ namespace stackjit {
 		  mOperandTypes(function->instructions()[instIndex].operandTypes()) {
 	}
 
-	StackFrameEntry StackFrame::getArgument(std::int64_t index) {
+	StackFrameEntry StackFrame::getArgument(std::size_t index) {
 		RegisterValue* argsStart = mBasePtr - 1;
 		return StackFrameEntry(argsStart - index, mFunction->def().parameters()[index]);
 	}
 
-	StackFrameEntry StackFrame::getLocal(std::int64_t index) {
+	StackFrameEntry StackFrame::getLocal(std::size_t index) {
 		RegisterValue* localsStart = mBasePtr - 1 - mFunction->def().numParams();
 		return StackFrameEntry(localsStart - index, mFunction->getLocal(index));
 	}
 
-	StackFrameEntry StackFrame::getStackOperand(std::int64_t index) {
+	StackFrameEntry StackFrame::getStackOperand(std::size_t index) {
 		RegisterValue* stackStart = mBasePtr - 1 - mFunction->def().numParams() - mFunction->numLocals();
 		return StackFrameEntry(stackStart - index, mOperandTypes[mOperandTypes.size() - 1 - index]);
 	}
